+++ src/components/UploadingJobsDisplay/UploadingJobsDisplay.js
@@ -6,7 +6,6 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { withRouter } from 'react-router-dom';
 import {
-  forEach,
   isEmpty,
   map,
   omit,
@@ -39,7 +38,6 @@ import { FileItem } from './components';
 import {
   xhrAddHeaders,
   generateSettingsLabel,
-  DEFAULT_TIMEOUT_BEFORE_FILE_DELETION,
   FILE_STATUSES,
 } from '../../utils';
 import * as API from '../../utils/upload';
@@ -57,6 +55,7 @@ export class UploadingJobsDisplay extends Component {
       block: PropTypes.func.isRequired,
       push: PropTypes.func.isRequired,
       replace: PropTypes.func.isRequired,
+      go: PropTypes.func,
     }).isRequired,
     location: PropTypes.oneOfType([
       PropTypes.shape({
@@ -66,10 +65,7 @@ export class UploadingJobsDisplay extends Component {
       }).isRequired,
       PropTypes.string.isRequired,
     ]),
-    timeoutBeforeFileDeletion: PropTypes.number, // milliseconds
   };
-
-  static defaultProps = { timeoutBeforeFileDeletion: DEFAULT_TIMEOUT_BEFORE_FILE_DELETION };
 
   static contextType = UploadingJobsContext;
 
@@ -79,6 +75,7 @@ export class UploadingJobsDisplay extends Component {
     files: {},
     hasLoaded: false,
     renderLeaveModal: false,
+    renderCancelUploadFileModal: false,
     recordsLoadingInProgress: false,
     JobProfilesComponent: null,
     actionMenuItems: ['run'],
@@ -87,7 +84,7 @@ export class UploadingJobsDisplay extends Component {
   async componentDidMount() {
     this.mounted = true;
     this.fileRemovalMap = {
-      [FILE_STATUSES.UPLOADED]: this.handleDeleteSuccessfullyUploadedFile,
+      [FILE_STATUSES.DELETING]: this.handleDeleteSuccessfullyUploadedFile,
       [FILE_STATUSES.ERROR]: this.deleteFileAPI,
       [FILE_STATUSES.ERROR_DEFINITION]: this.deleteFileFromState,
     };
@@ -101,13 +98,11 @@ export class UploadingJobsDisplay extends Component {
   componentWillUnmount() {
     this.mounted = false;
 
-    this.cancelFileRemovals();
     this.resetPageLeaveHandler();
   }
 
   calloutRef = createRef();
-
-  deleteFileTimeouts = {};
+  selectedFile = null;
 
   mapFilesToState() {
     const {
@@ -175,11 +170,6 @@ export class UploadingJobsDisplay extends Component {
     const { files } = this.state;
 
     return !this.isSnapshotMode && some(files, file => file.status === FILE_STATUSES.UPLOADING);
-  }
-
-  cancelFileRemovals() {
-    forEach(this.deleteFileTimeouts, clearTimeout);
-    this.deleteFileTimeouts = {};
   }
 
   renderSnapshotData() {
@@ -338,22 +328,21 @@ export class UploadingJobsDisplay extends Component {
     this.updateFileState(fileKey, { status: FILE_STATUSES.ERROR });
   };
 
-  handleUndoDeleteFile = fileKey => {
-    clearTimeout(this.deleteFileTimeouts[fileKey]);
+  handleDeleteFile = () => {
+    const {
+      key,
+      status,
+    } = get(this.state, ['files', this.selectedFile]);
 
-    this.updateFileState(fileKey, { status: FILE_STATUSES.UPLOADED });
-  };
-
-  handleDeleteFile = (fileKey, fileStatus) => {
-    const deleteFile = this.fileRemovalMap[fileStatus];
-
+    const deleteFile = this.fileRemovalMap[status];
     if (deleteFile) {
-      deleteFile(fileKey, fileStatus);
+      deleteFile(key, status);
     }
   };
 
   deleteFileAPI = async (fileKey, fileStatus) => {
     const { files: { [fileKey]: fileMeta } } = this.state;
+    const { history } = this.props;
 
     this.updateFileState(fileKey, { loading: true });
 
@@ -385,10 +374,15 @@ export class UploadingJobsDisplay extends Component {
       console.error(error); // eslint-disable-line no-console
     }
 
+    this.setState({ renderCancelUploadFileModal: false });
+
     const lastFileDeleted = isEmpty(get(this.state, ['files'], {}));
 
     if (lastFileDeleted) {
-      await this.updateJobProfilesComponent();
+      history.push({
+        pathname: '/data-import',
+        search: '',
+      });
     }
   };
 
@@ -415,13 +409,7 @@ export class UploadingJobsDisplay extends Component {
   });
 
   handleDeleteSuccessfullyUploadedFile = (fileKey, fileStatus) => {
-    const { timeoutBeforeFileDeletion } = this.props;
-
-    this.deleteFileTimeouts[fileKey] = setTimeout(() => {
-      this.deleteFileAPI(fileKey, fileStatus);
-    }, timeoutBeforeFileDeletion);
-
-    this.updateFileState(fileKey, { status: FILE_STATUSES.DELETING });
+    this.deleteFileAPI(fileKey, fileStatus);
   };
 
   renderFiles() {
@@ -461,8 +449,8 @@ export class UploadingJobsDisplay extends Component {
           uploadedValue={uploadedValue}
           errorMsgTranslationID={errorMsgTranslationID}
           uploadedDate={uploadedDate}
+          onCancel={this.openCancelUploadModal}
           onDelete={this.handleDeleteFile}
-          onUndoDelete={this.handleUndoDeleteFile}
         />
       );
     });
@@ -479,6 +467,17 @@ export class UploadingJobsDisplay extends Component {
   closeModal = () => {
     this.setState({ renderLeaveModal: false });
   };
+
+  openCancelUploadModal = (fileKey) => {
+    this.setState({ renderCancelUploadFileModal: true });
+    this.updateFileState(fileKey, { status: FILE_STATUSES.DELETING });
+    this.selectedFile = fileKey;
+  }
+
+  closeCancelUploadModal = () => {
+    this.updateFileState(this.selectedFile, { status: FILE_STATUSES.UPLOADED });
+    this.setState({ renderCancelUploadFileModal: false });
+  }
 
   /** @returns {Promise<string[]>} */
   async getDataTypes() {
@@ -518,6 +517,7 @@ export class UploadingJobsDisplay extends Component {
     const {
       hasLoaded,
       renderLeaveModal,
+      renderCancelUploadFileModal,
       JobProfilesComponent,
       actionMenuItems,
     } = this.state;
@@ -571,6 +571,15 @@ export class UploadingJobsDisplay extends Component {
               cancelLabel={<FormattedMessage id="ui-data-import.modal.leavePage.cancel" />}
               onConfirm={this.closeModal}
               onCancel={this.continue}
+            />
+            <ConfirmationModal
+              open={renderCancelUploadFileModal}
+              heading={<FormattedMessage id="ui-data-import.modal.cancelUpload.header" />}
+              message={<FormattedMessage id="ui-data-import.modal.cancelUpload.message" />}
+              confirmLabel={<FormattedMessage id="ui-data-import.modal.cancelUpload.confirm" />}
+              cancelLabel={<FormattedMessage id="ui-data-import.modal.cancelUpload.cancel" />}
+              onConfirm={this.handleDeleteFile}
+              onCancel={this.closeCancelUploadModal}
             />
           </Pane>
         </div>